+++ src/commands/commandList.js
@@ -1,77 +1,24 @@
-const { PrivateKeyService } = require('../services/PrivateKeyService')
-const OrderService = require('../services/OrderService')
-const logger = require('../logger')
+const ApproveCommandBuilder = require('../factories/commands/ApproveCommandBuilder')
+const CreateWalletCommandBuilder = require('../factories/commands/CreateWalletCommandBuilder')
+const DepositEthCommandBuilder = require('../factories/commands/DepositEthCommandBuilder')
+const DepositTokenCommandBuilder = require('../factories/commands/DepositTokenCommandBuilder')
+const GetAddressCommandBuilder = require('../factories/commands/GetAddressCommandBuilder')
+const GetAllowanceCommandBuilder = require('../factories/commands/GetAllowanceCommandBuilder')
+const GetBalanceCommandBuilder = require('../factories/commands/GetBalanceCommandBuilder')
+const OrderCancelCommandBuilder = require('../factories/commands/OrderCancelCommandBuilder')
+const OrderSignCommandBuilder = require('../factories/commands/OrderSignCommandBuilder')
+const WithdrawCommandBuilder = require('../factories/commands/WithdrawCommandBuilder')
 
-const ApproveCommand = require('../commands/ApproveCommand')
-const ApproveCommandValidator = require('../validators/ApproveCommandValidator')
-const CreateWalletCommand = require('../commands/CreateWalletCommand')
-const CreateWalletCommandValidator = require('../validators/CreateWalletCommandValidator')
-const DepositEthCommand = require('./DepositEthCommand')
-const DepositEthCommandValidator = require('../validators/DepositEthCommandValidator')
-const DepositTokenCommand = require('./DepositTokenCommand')
-const DepositTokenCommandValidator = require('../validators/DepositTokenCommandValidator')
-const GetAddressCommand = require('../commands/GetAddressCommand')
-const GetAddressCommandValidator = require('../validators/GetAddressCommandValidator')
-const GetAllowanceCommand = require('../commands/GetAllowanceCommand')
-const GetAllowanceCommandValidator = require('../validators/GetAllowanceCommandValidator')
-const GetBalanceCommand = require('../commands/GetBalanceCommand')
-const GetBalanceCommandValidator = require('../validators/GetBalanceCommandValidator')
-const OrderCancelCommand = require('../commands/OrderCancelCommand')
-const OrderCancelCommandValidator = require('../validators/OrderCancelCommandValidator')
-const OrderSignCommand = require('../commands/OrderSignCommand')
-const OrderSignCommandValidator = require('../validators/OrderSignCommandValidator')
-const OrderSignerHelper = require('../helpers/OrderSignerHelper')
-const PrivateKeyValidator = require('../validators/PrivateKeyValidator')
-const TradingWalletServiceBuilder = require('../factories/TradingWalletServiceBuilder')
-const WithdrawCommand = require('../commands/WithdrawCommand')
-const WithdrawCommandValidator = require('../validators/WithdrawCommandValidator')
-
-const approveCommandValidator = new ApproveCommandValidator(logger)
-const createWalletCommandValidator = new CreateWalletCommandValidator(logger)
-const depositEthCommandValidator = new DepositEthCommandValidator(logger)
-const depositTokenCommandValidator = new DepositTokenCommandValidator(logger)
-const getAddressCommandValidator = new GetAddressCommandValidator(logger)
-const getAllowanceCommandValidator = new GetAllowanceCommandValidator(logger)
-const getBalanceCommandValidator = new GetBalanceCommandValidator(logger)
-const orderCancelCommandValidator = new OrderCancelCommandValidator(logger)
-const orderSignCommandValidator = new OrderSignCommandValidator(logger)
-const ordersignerHelper = new OrderSignerHelper(logger)
-const orderService = new OrderService(logger)
-const privateKeyService = new PrivateKeyService(logger)
-const privateKeyValidator = new PrivateKeyValidator(logger)
-const withdrawCommandValidator = new WithdrawCommandValidator(logger)
-
-const tradingWalletService = TradingWalletServiceBuilder.build()
-
-const getAddressCommand = new GetAddressCommand(logger, tradingWalletService,
-  getAddressCommandValidator, privateKeyService, privateKeyValidator)
-
-const getAllowanceCommand = new GetAllowanceCommand(logger,
-  getAllowanceCommandValidator, privateKeyService, privateKeyValidator)
-
-const getBalanceCommand = new GetBalanceCommand(logger, tradingWalletService,
-  getBalanceCommandValidator, privateKeyService, privateKeyValidator)
-
-const createWalletCommand = new CreateWalletCommand(logger, getAllowanceCommandValidator,
-  createWalletCommandValidator, privateKeyService, privateKeyValidator)
-
-const depositEthCommand = new DepositEthCommand(logger, tradingWalletService,
-  depositEthCommandValidator, privateKeyService, privateKeyValidator)
-
-const depositTokenCommand = new DepositTokenCommand(logger, tradingWalletService,
-  depositTokenCommandValidator, privateKeyService, privateKeyValidator)
-
-const signCommand = new OrderSignCommand(logger, ordersignerHelper, orderSignCommandValidator,
-  privateKeyService, privateKeyValidator)
-
-const withdrawCommand = new WithdrawCommand(logger, tradingWalletService, withdrawCommandValidator,
-  privateKeyService, privateKeyValidator)
-
-const approveCommand = new ApproveCommand(logger, approveCommandValidator,
-  privateKeyService, privateKeyValidator)
-
-const orderCancelCommand = new OrderCancelCommand(logger, orderService, orderCancelCommandValidator,
-  privateKeyService, privateKeyValidator)
+const approveCommand = ApproveCommandBuilder.build()
+const createWalletCommand = CreateWalletCommandBuilder.build()
+const depositEthCommand = DepositEthCommandBuilder.build()
+const depositTokenCommand = DepositTokenCommandBuilder.build()
+const getAddressCommand = GetAddressCommandBuilder.build()
+const getAllowanceCommand = GetAllowanceCommandBuilder.build()
+const getBalanceCommand = GetBalanceCommandBuilder.build()
+const orderCancelCommand = OrderCancelCommandBuilder.build()
+const signCommand = OrderSignCommandBuilder.build()
+const withdrawCommand = WithdrawCommandBuilder.build()
 
 module.exports = {
   approveCommand,